Move program_clauses_that_could_match to chalk-solve
