+++ Cargo.toml
@@ -24,7 +24,6 @@ salsa = "0.10.0"
 serde = "1.0"
 serde_derive = "1.0"
 stacker = "0.1.5"
-rustc-hash = { version = "1.0.0" }
 
 [dependencies.chalk-parse]
 version = "0.1.0"
