+++ chalk-solve/Cargo.toml
@@ -15,6 +15,7 @@ ena = "0.10.1"
 failure = "0.1"
 itertools = "0.7.8"
 petgraph = "0.4.13"
+rustc-hash = { version = "1.0.0" }
 
 [dependencies.chalk-macros]
 version = "0.1.0"
